requirements/local.txt
asgiref
Django
django-model-utils
djangorestframework
Pillow
psycopg2
sqlparse
Unipath
firebase-admin